Slow

Documentation for Slow.

Slow.@slow

Slow.@slowdef

Slow.@slowtest

Slow.@slow — Macro

@slow

Call a slow version of a function that was defined with @slowdef. ```

Slow.@slowdef — Macro

@slowdef

Define a slow version of a function which can be called with @slow.

Slow.@slowtest — Macro

@slowtest

Shortcut for @test (@slow func(args...)) == func(args...).