Slow

Documentation for Slow. This package attempts to address a tension between understandability and efficiency in scientific computing. A naive Julia implementation of a calculation is often very readable, and resembles published equations or pseudocode. However, an optimized implementation that uses knowledge of how modern CPUs work can be much faster – but it's usually more verbose.

This package exports @slowdef to define a slower, naive implementation of a function, and call it with @slow. Here's an example, where we implement a slow version of a function, and a fast version.

using Slow

# fake naive implementation
@slowdef function f(x)
    sleep(1)  
    return sin(x)
end

# fake fast implementation
function f(x)
    return sin(x)
end

@time @slow f(1.0)
@time f(1.0)

The function definition func(args...) prefaced by @slowdef is replaced with signature func(::Slow.SlowImplementation, args...). Nested use of func can now be toggled between the slow and fast implementations.

h(x) = f(x)^2 + cos(x)^2

@time @slow h(1.0)
@time h(1.0)

1.002132 seconds (18 allocations: 464 bytes)
0.000000 seconds

Note that the allocations here arise from the use of @time, not @slow which only operates before compilation.

Single Function Selection

By default, @slow slows every function involved in the expression which has a slow implementation in the caller's module. It can sometimes be desirable to slow down a specific function. This is achieved by providing a function before the expression to be evaluated by @slow.

julia> @slowdef function s(x)
           println("slow s") 
           return sin(x)
       end
       s(x) = sin(x)
       
       @slowdef function c(x)
           println("slow c") 
           return cos(x)
       end
       c(x) = cos(x)

       h(x) = s(x)^2 + c(x)^2

julia> @slow s h(0.5)
slow s
1.0

julia> @slow c h(0.5)
slow c
1.0

julia> @slow h(0.5)
slow s
slow c
1.0

API

Slow.@slow

Slow.@slowdef

Slow.@slowtest

Slow.@slow — Macro

@slow

Call a slow version of a function that was defined with @slowdef. ```

Slow.@slowdef — Macro

@slowdef

Define a slow version of a function which can be called with @slow.

Slow.@slowtest — Macro

@slowtest

Shortcut for @test (@slow func(args...)) == func(args...).