ReferenceImplementations.jl

Documentation for ReferenceImplementations.jl.

This package exports @refimpl to define a reference implementation of a function, and call it (even when it's buried in some other function) using the @refimpl macro. It does this by performing a Cassette pass on every top-level function call in the expression provided to the macro.

Why?

I often write two versions of a function,

V1: Naive implementation. Since Julia is so expressive, this implementation is usually short and resembles the published equations or pseudocode.

V2: Optimized implementation. This version is written for a computer, i.e. ⊂ { exploits symmetries, reuses allocated memory, hits the cache in a friendly way, reorders calculations for SIMD, divides the work with threads, precomputes parts, caches intermediate expressions, ... }.

V1 is a reference implementation that is correct, easy to understand, and easy to extend. V2 is the implementation exported in your package and it's often much faster, but complicated and verbose. Julia sometimes allows you to use abstractions such that V1 ≈ V2, but this is not always possible. ReferenceImplementations.jl lets you keep both, and toggle them even when they're deeply nested.

Usage

Here's an example, where we implement a reference implementation of a function, and a fast version.

using ReferenceImplementations

# fake naive implementation
@refimpl function f(x)
    println("ref f")
    return sin(x)
end

# fake fast implementation
function f(x)
    println("f")
    return sin(x)
end

f(0.0)

# output

f
0.0

Running f(0.0) just uses the definition we gave for it. However, @refimpl f(0.0) will go to the reference implementation.

@refimpl f(0.0)

# output

ref f
0.0

The function definition func(args...), when prefaced by @refimpl, instead defines a function with signature func(::ReferenceImplementations.RefImpl, args...). Use of func can now be be substituted with the reference implementation, even when the calls are nested in other functions.

julia> h(x) = f(x)^2 + cos(x)^2
h (generic function with 1 method)

julia> h(1.0)
f
1.0

julia> @refimpl h(1.0)
ref f
0.0

Single Function Selection

By default, @refimpl replaces every call in the expression which has a defined reference implementation. It can sometimes be desirable to use the reference implementation of a specific function. This is achieved by providing a function before the expression to be evaluated by @refimpl (func) (expr).

julia> @refimpl s(x) = begin println("ref s"); return sin(x) end
       s(x) = sin(x)
       @refimpl c(x) = begin println("ref c"); return cos(x) end
       c(x) = cos(x)
       h(x) = s(x)^2 + c(x)^2;

julia> @refimpl s h(0.5)
ref s
1.0

julia> @refimpl c h(0.5)
ref c
1.0

julia> @refimpl h(0.5)
ref s
ref c
1.0

API

ReferenceImplementations.@refimpl

ReferenceImplementations.@refimpl — Macro

@refimpl (method definition)
@refimpl (expression)
@refimpl (function name) (expression)

If prefacing a function definition, defines a reference implementation for that function.

If prefacing an expression, calls reference implementations of methods defined with @refimpl.

Preface an expression performs a Cassette pass on every top-level function in the expression, recursively looking for methods with @refimpl implementations. If a function is passed before the expression (separated by a space), only that method is switched with its reference implementation.

Examples

Calling @refimpl on an expression calls every method with a reference implementation in the nested sequence of calls for that expression.

using ReferenceImplementations
@refimpl mysin(x) = begin println("ref mysin"); return sin(x) end
mysin(x) = begin println("mysin"); return sin(x) end

# call the reference implementation
@refimpl mysin(0.)  # prints "ref mysin"
mysin(0.)        # prints "mysin"

This works for @refimpl functions that are nested inside other functions in the expression.

@refimpl f(x) = begin println("ref f"); return mysin(x)^2 end
f(x) = begin println("f"); return mysin(x)^2 end

# call the reference implementation
@refimpl f(0.)  # prints "ref f", "ref mysin"
f(0.)        # prints "f", "mysin"

You can target individual functions to be replaced with their reference implementation by passing that function after @refimpl.

@refimpl mysin f(0.)  # prints "fast f", "ref mysin"
@refimpl f f(0.)  # prints "ref f", "fast mysin"